HotkeyManager: Trigger hotkeys by specificity
More specific hotkeys get priority over less specific ones. EG:

LeftControl+Y, LeftControl+X, and X are all bound.

The user is pressing LeftControl+X+Y.

In the old system all of these hotkeys would be triggered at
once. In the new one it wouldn't trigger hotkey X.

In the new one it will first gather the highest specificity
among the hotkeys (In this case LeftControl+X nad LeftControl+Y
both have a specificity of 2, while X has a specificity of 1)
and execute them.

So in other words, LeftControl+X has a specificity of 2 and X only
has a specificity of 1, so X won't be triggered.